Backfill - Plex version 1.32.2.7002-86cfcc10c

+++ plex-ffmpeg-source/PlexTranscoder/libavformat/matroskaenc.c
@@ -3208,7 +3208,7 @@ static int mkv_check_bitstream(AVFormatContext *s, AVStream *st,
         ret = ff_stream_add_bitstream_filter(st, "vp9_superframe", NULL);
     } else if (st->codecpar->codec_id == AV_CODEC_ID_H264) {
         if (!st->codecpar->extradata_size) {
-            int side_data_size;
+            size_t side_data_size;
             uint8_t *side_data = av_packet_get_side_data(pkt, AV_PKT_DATA_NEW_EXTRADATA,
                                                          &side_data_size);
